chore: upgrade all crates, migrate Fjall 2→3, remove dead data structures
- Fjall 2.0 → 3.1.3 (Database/Keyspace hierarchy, KeyspaceCreateOptions, OwnedWriteBatch)
- axum 0.7 → 0.8, tower-http 0.5 → 0.6, reqwest 0.11 → 0.12
- petgraph 0.6 → 0.8, strum 0.24 → 0.28, thiserror 1 → 2
- rstest 0.18 → 0.26, criterion 0.5 → 0.8, async-nats 0.33 → 0.47
- Remove dead deps: serde_yaml, nom, async-nats
- Remove 27 unused data structure modules from vo-types (~11K lines)
- Fix reanimator runtime nesting bug (block_on inside async runtime)
- Fix spawn supervisor Failed→Spawn respawn path
- 5,275 tests passing across all crates

+++ crates/vo-worker/src/pool/hash_ring.rs
@@ -151,13 +151,13 @@ impl HashRing {
         let mut hasher = CrcHasher::new();
         pool_id.as_str().hash(&mut hasher);
         virtual_index.hash(&mut hasher);
-        hasher.finish() as u64
+        hasher.finish()
     }
 
     fn hash_key<K: Hash>(&self, key: &K) -> u64 {
         let mut hasher = CrcHasher::new();
         key.hash(&mut hasher);
-        hasher.finish() as u64
+        hasher.finish()
     }
 }
